Making new phenology 2024 file 12.12.24 Merging two files that had the last DENU phenology survey data for 2024:
Also removing all the other sheets in the file (prior demography sheets and blank datasheet)

12.13.24
put a "1" in florivory for A.98.NA on 6/19 and A.201.GD  because note said hole in flower.

12.14.24
deleted "delete" column after consulting with Fatima who says it was notes to the data-takers about which plants to skip because they were done.
DENU.B.230.BS added uniqueID on 6/26 - cell was blank
Removing blank lines (no data taken for those plants on those days).  Up through 7/10 because I just figured out that blank lines that day aren't correct.
Fixes to 7/10 coming in separate commit.